requirements.txt
discord.py
pymongo
dsnpython
